requirements.txt
beautifulsoup4==4.12.3
PyYAML==6.0.1
Requests==2.32.3
selenium==4.26.1
